ai-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.335
langchain-community==0.0.10
tavily-python==0.3.0
pydantic==2.5.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
requests==2.31.0
